Eliminated flicker on update and in images
Changed some of the matrix hardware flags to reduce flickering. Also added a field variable to ensure the matrix only updates the display when something changes, instead of every second like it used to. This way it isn't swapping out an image with the same image and doing overhead calculations (which caused flickering in both scenarios).