Made nicer UI, added better color control, background images
There is now a UI for colors, you click on the color button and it will come up with a menu to select, or you can still type a command. You can also now change the colors of the weather, time, and date. If you use custom colors the night/day switch is currently disabled (will be fixed later)

You can also now add images to be used as a background or full screen image (full screen meaning no clock).